+++ lib/redis_queued_locks/acquier/lock_info.rb
@@ -11,7 +11,7 @@ module RedisQueuedLocks::Acquier::LockInfo
     #   - result format: {
     #     lock_key: "rql:lock:your_lockname", # acquired lock key
     #     acq_id: "rql:acq:process_id/thread_id", # lock acquier identifier
-    #     ts: 123456789, # <locked at> time stamp (epoch)
+    #     ts: 123456789.2649841, # <locked at> time stamp (epoch, seconds.microseconds)
     #     ini_ttl: 123456789, # initial lock key ttl (milliseconds),
     #     rem_ttl: 123456789, # remaining lock key ttl (milliseconds)
     #   }
@@ -46,7 +46,7 @@ module RedisQueuedLocks::Acquier::LockInfo
           {
             lock_key: lock_key,
             acq_id: hget_cmd_res['acq_id'],
-            ts: Integer(hget_cmd_res['ts']),
+            ts: Float(hget_cmd_res['ts']),
             ini_ttl: Integer(hget_cmd_res['ini_ttl']),
             rem_ttl: ((pttl_cmd_res == -1) ? Infinity : pttl_cmd_res)
           }
